ec: Add support for reading the LID switch
Add a function for reading the lid switch from the EC via the memmap. This
can be used to read the lid switch, for example, since if it is only
connected to the EC there is no way to use the AP GPIOs.

BUG=chrome-os-partner:53208
BRANCH=none

TEST=with the rest of the patches applied, on a Gru device, while
     observing the AP console:

  - reboot the device into linux login prompt

  - reboot the device again, close the lid once the developer mode
    screen is displayed. Observe the device stop booting, printing on
    the console
VbAudioOpen() - note count 1
Lid is closed.
VbBootDeveloper() - shutdown requested!

Change-Id: I705cac1a5ac781d4f382c7079babfb279cdc8627
Signed-off-by: Simon Glass <sjg@chromium.org>
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/353940
Reviewed-by: Shawn N <shawnn@chromium.org>