usb_eth: Remove invalid free of GenericUsbDevice.dev_data
The USB ethernet driver framework currently frees a pointer on USB
unplug/shutdown that had never been initialized. This might have stemmed
from confusion between the NetDevice and GenericUsbDevice objects...
both provide a dev_data member pointer for use by specific driver
instantiations, but the latter's is not actually used by this driver.
(The former's is just used as a back-pointer to the GenericUsbDevice,
which already gets freed by the generic USB stack and doesn't need to be
freed by the driver.)

BRANCH=None
BUG=None
TEST=With improved memory allocator error messages, I no longer see an
out-of-heap free(0x840) on Kevin.

Change-Id: I30adc14659f9cec7c342df760ebef83ff422f9ae
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/366859
Reviewed-by: Aaron Durbin <adurbin@chromium.org>