cros_ec: Improve SPI polling behavior after reboot to RW
Our error handling during EC SPI communication is pretty gnarly: we just
wait for up to a whole second to see the start-of-frame byte (0xec)
appear and ignore everything else. The EC actually clocks out a
meaningful status byte whenever CS is asserted which can tell us what
state of host command processing it is in and whether it has already
encountered a fatal error for the current command. This patch makes use
of that for early abort in known-fatal conditions and keeps the timeout
much smaller if the EC doesn't tell us that it has actually received our
full command and started to process it.

The latter is mostly useful for the case of polling the EC after a
sysjump to RW. If the EC doesn't bring up its SPI interface within the
initial (mandated) 50ms delay (which is normal for some chips like the
npcx), we'll keep polling it in much smaller intervals to avoid wasting
too much time. This also necessitates some more sophisticated handling
of error output so we don't spam the console for those expected timeouts
while we're waiting for the EC to come back up.

BRANCH=None
BUG=chrome-os-partner:56254
TEST=Booted Kevin, forced a software sync, made sure everything still
works and the jump to RW is now much faster.

Change-Id: I85a71fd6c361ae7236736580d2ca52c652f45dc5
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/371642
Reviewed-by: Shawn N <shawnn@chromium.org>