cros_ec: Don't print to console on ec_test() error
ec_test() is a function intended to be used to poll whether the EC is
ready to communicate. Since this implies that the EC might not be ready
for a while, it is expected that this function might fail for a few
tries and it should not spam the console every time that happens
(instead leave error handling to the caller). This patch removes such a
console print that had previously never been executed due to a bug when
checking the return value of ec_command(), which was fixed in CL:379215.

BRANCH=None
BUG=None
TEST=Booted Kevin, no longer saw a bunch of "ec_command() returned
error" spam after EC jump to RW.

Change-Id: I540d590fcd37193542040f1f444430fb11cb1546
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/380599
Reviewed-by: David Hendricks <dhendrix@chromium.org>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>