drivers/tpm: Move probe handling into specific driver
There is some probe code in the high level i2c tpm driver
that is attempting to write zero to the access register with
the basic i2c_writeb() methods which are not appropriate for
all TPMs like the CR50 which expect accesses to be delayed
properly.

BUG=chrome-os-partner:53336
BRANCH=none
TEST=verify with i2c debug enabled that no unexpected i2c
transactions to the TPM are seen.

Change-Id: I193a8266c50a18fb2487c8e6b67855df6f814a9b
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/382730
Reviewed-by: Aaron Durbin <adurbin@chromium.org>