oak: Make CONFIG_BOARD_ID_ADJUSTMENT overridable in defconfig
In its infinite wisdom, Kconfig decides to only honor pre-existing
values in your .config during make oldconfig if that config option has a
help text... otherwise, it will just ignore the preset value and use its
default. For Oak's CONFIG_BOARD_ID_ADJUSTMENT we decided to default to 7
(which is true for anything *but* Oak to date) and override the value
for Oak itself in its defconfig. However, apparently I never actually
tested that and the missing help text causes the override to not be
applied.

BRANCH=oak
BUG=None
TEST=Oak boots again on ToT.

Change-Id: I6f6fe0b5e0f17944cd62be9b0b3da7b452a9a592
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/386296
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>