tpm: i2c: cr50: Add interrupt support to driver
Add support for reading interrupt status to the TPM driver
and use that instead of a fixed delay while waiting for the
TPM to be ready.

This also sync's the driver with coreboot function names
and comments in order to keep the two similar.

In order to do the interrupt read the board provides a method
that will be called by the tpm driver.  For reef this uses
the apollolake GPE interface.

BUG=chrome-os-partner:56663
BRANCH=none
TEST=boot on reef with TPM enabled

Change-Id: If34652fd9322bc5bbcaae62b4cc98efc8a521c88
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/389240
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>