apollolake: allow for the PMC device's config space to be hidden
The current PCH code assumes the config space for the device
containing the ACPI bar is always available. On the apollolake
boards that's not always true. To work around this provide an
option that indicates the config space for the pci device may
not be present. The apollolake code then checks for this condition.

BUG=chrome-os-partner:58551
BRANCH=None
TEST=Booted system with new EC RW firmware. Noted the device shutdown
     correctly instead of declaring 'Power off is not implemented for
     this chipset. Halting the CPU.'

Change-Id: I3808d0b8700179df24d221266ce77cd83b86d460
Signed-off-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/398159
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Reviewed-by: Duncan Laurie <dlaurie@google.com>