Makefile: Fix default for LIBPAYLOAD_DIR
We're usually passing LIBPAYLOAD_DIR to depthcharge's Makefile, but it's
also possible to omit that parameter and have it default to the location
that libpayload would be in if depthcharge's source is located under
coreboot/payloads/.

Unfortunately, it looks like that hasn't been working right for a while.
We're trying to source xcompile from libpayload before we're setting
that default, and we're later passing the derived location of lpgcc to
vboot without handling the possibility that LIBPAYLOAD_DIR was a
relative path. This patch resolves both issues. Since LIBPAYLOAD_DIR is
usually not set for the Kconfig pass of the Makefile, move the code down
into the respective ifeq section.

BRANCH=None
BUG=None
TEST=Build and booted depthcharge for Kevin without emerge.

Change-Id: Icec7262c7a0748944b0862fe7833a617a7ab567c
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/404903