sound: Add optional disable callback to SoundRouteComponent
Depthcharge likes to always just turn things on and forget about them,
including the sound route components. This is usually fine, but
occasionally we might have components that really need to be disabled
again after we're done playing audio to ensure the thermal integrity
of our speakers. This patch adds an optional callback that can be
used for that purpose.

BRANCH=gru
BUG=chrome-os-partner:59038
TEST=Booted Kevin, listened to dev mode beep.

Change-Id: I7fbeadd6a1ef660edc9f5223e8e223b4416ea41c
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/405869
Tested-by: Xing Zheng <zhengxing@rock-chips.com>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>