eve: Enable sound path for dev mode beep
Enable the GPIO I2S interface and the Max98927 codec driver
to provide a path for dev mode beep.  This uses the same
GPIO I2S driver that was used on skylake boards with the
same configuration.

Use GPIO D22 to control the I2S isolation buffer chips and
have it reset upon exit to enable the normal I2S bus.

BUG=chrome-os-partner:58666
BRANCH=none
TEST=boot on eve in dev mode and listen for beep

Change-Id: I49c83e150ffaea36648434e2b07b3c352a6d02ca
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/418903
Reviewed-by: Aaron Durbin <adurbin@chromium.org>