tpm2: handle failures more gracefully
When trying to bring up a device with a malfunctioning TPM2 chip, the
driver currently gets stuck waiting for SPI flow control, causing
bricked devices.

This patch puts a 100 ms cap on the waiting time - this should be
enough even for a longest NVRAM save operation which could be under
way on the TPM device.

BRANCH=gru
BUG=chrome-os-partner:59807
TEST=with a matching change in coreboot, now a gru with corrupted SPI
     TPM comes up to the recovery screen (it was not showing signs of
     life before this change).

Change-Id: I6fc73b2ac76ccb3bac01f52707f5016505713fca
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/421174
Reviewed-by: Aaron Durbin <adurbin@chromium.org>