tpm: add tpm state report plumbing
This patch makes use of a new vendor command supported by CR50, which
allows to receive some internal state information from the TPM.

The structure of the information returned by the TPM is defined in the
EC codebase and needs to be kept in sync manually.

Presently the lockout counter values and information about TPM
failures (if any) are returned.

BRANCH=none
BUG=chrome-os-partner:60555
TEST=with the rest of the patches applied to depthcharge and ec trees
      observed tpm state information reported on the recovery screen
      in response to hitting the tab key

Change-Id: I5d7ec39edc25f954ddb67220f7319f4a3d6741a3
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/430190
Reviewed-by: Aaron Durbin <adurbin@chromium.org>