Ensure that vboot is built in the expected directory
At present the vboot Makefile is called with a relative path, but since
it does not (appear to) run in the current directory, it can output files
into the wrong directory.

Pass the vboot Makefile an absolute path to resolve this problem. This
allows us (for example) to build depthcharge with a non-absolute directory
path.

BUG=chromium:685706
BRANCH=none
TEST=emerge-reef --nodeps depthcharge; see that it builds OK

Signed-off-by: Simon Glass <sjg@chromium.org>

Change-Id: I5e363acea85608ff4f3e30550a578c9ae875f734
Reviewed-on: https://chromium-review.googlesource.com/433342
Commit-Ready: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>