rk3399_sdhci: Only power-cycle the PHY for last frequency bump
The Arasan SHDCI controller seems to behave slightly differently on
every board. We've now seen a device that doesn't work reliably with the
existing code (the first access after switching to HS 52MHz fails). It
seems that *not* power-cycling the PHY for that frequency change solves
the problem. We don't really know why, but we only ever really needed
the power-cycle to make the final frequency change (to HS400es 200MHz)
work well, so let's go with that for now.

BRANCH=gru
BUG=chrome-os-partner:63025
TEST=Reboot stress test on three Kevins (one showing the original bug).

Change-Id: Ic4b14c90d131b654caddd259f08a156cc94fb8b7
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/444031