cros_ec: Use a dedicated structure for mkbp state to stay within bounds
The cros_ec_scan_keyboard function is expected to fill an array of
keyscan cols, provided by the EC as a key matrix.

This was done by casting the key_matrix element of the
ec_response_get_next_data structure as a cros_ec_keyscan structure, that
initially only contained a similar key matrix, making the cast somewhat
valid. However, the buttons element was added to that structure, which
increased its size. As ec_command is called with the size of that
structure and copies that number of bytes to the provided location, the
added bytes ended up copying data out of the bounds of the initial key
matrix, leading to unpredictible results.

On a specific test run with a big, this would change the first byte of
the (unrelated) sound_ops variable, causing a prefetch error exception.

Passing the key matrix (instead of the whole cros_ec_keyscan structure)
from a proper cros_ec_keyscan structure to ec_command (with the right
size) resolves the issue. The data is then copied over to the
ec_response_get_next_data structure as needed for the rest of the code
to work normally.

BUG=none
BRANCH=none
TEST=boot on big: keyboard works and it doesn't hang anymore

Change-Id: If77a721c10b179e0e76cfcb706ea972fbf5646b9
Signed-off-by: Paul Kocialkowski <contact@paulk.fr>
Reviewed-on: https://chromium-review.googlesource.com/446096
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Shawn N <shawnn@chromium.org>