drivers/input/mkbp: Add a new option DRIVER_INPUT_MKBP_NO_INTERRUPT
In case of x86 Chromebooks, there is no interrupt line coming from the
EC to the AP. In such cases, it is better to read host event first and
see if there is a pending MKBP event. If yes, only then try to read
the MKBP event fifo. Config option DRIVER_INPUT_MKBP_NO_INTERRUPT
enables this behavior by default for x86. Additionally, it can be used
by any board that does not have a dedicated interrupt line from EC to
AP.

BUG=b:35774934
BRANCH=None
TEST=Verified that depthcharge no longer asks EC for MKBP events if
none are queued. Also, gets rid of all the spew of messages reporting
failure to read MKBP event since it was not queued.

Change-Id: I1ad732a88c6c1402eaabe391fd799efe836d1cce
Signed-off-by: Furquan Shaikh <furquan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/452672
Reviewed-by: Aaron Durbin <adurbin@chromium.org>