COMMIT-QUEUE: Configure pre-CQ for more coverage
The pre-CQ doesn't really do its job very well for firmware at the
moment, since it only runs a couple of boards and misses a lot of our
mainboard-specific code. This patch is an attempt to improve this by
manually specifying a list of recent boards that are most likely to
break from new patches. This list will have to be continuously updated
as we are adding new baseboards.

BRANCH=none
BUG=chromium:701504
TEST=none

Change-Id: Ie7b1cf0623a78e1d817b51bbef9c0b120a72682c
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/454881
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Reviewed-by: Stefan Reinauer <reinauer@google.com>