rk3399_sdhci: Reintroduce PHY power-cycling at 52MHz
It seems like there are some other Kevin DUTs for which *not* power
cycling the PHY at 52MHz is a bad idea. They get random MMC errors
shortly afterwards (e.g. when setting the bus width or when trying to
switch to HS400ES).

This patch enables the PHY power-cycle for that transition again, which
is fine since we're now correctly setting the DLL freqsel which also
seemed to help with our previous problem. This sequence should be nearly
identical (save for duplicate transitions and timing) with the kernel
sequence, so let's hope that this finally solves the issue for *all*
Kevins.

BRANCH=gru
BUG=b:35648440
TEST=Ran a tight boot loop on the affected DUT without seeing it enter
recovery mode.

Change-Id: Ie84ee6b2546f3f6ae8a4a50684c82ffc59bc5748
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/462039
Reviewed-by: Douglas Anderson <dianders@chromium.org>