multiboot: Add support for bootdata
The multiboot specification does not provide a way to pass some
useful information to the kernel.  The bootdata specification
provides structures for passing framebuffer, serial, and
lastlog/ramoops to the kernel.

Implement support for these data strucutres and fill them out
if they are found in the sysinfo populated from coreboot tables.

BUG=b:38040849
BRANCH=eve
TEST=manual testing with bootdata format kernel

Change-Id: I2e6a5900684aa35dac04cb48bf79d7e74a790d46
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Commit-Id: 1a2fcfcd131534c267eeba8af221e46e0bf9435c
Original-Change-Id: I18c0861e808ce5ea2eb55366f092841c55c14970
Original-Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Reviewed-on: https://chromium-review.googlesource.com/547402
Original-Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/565941