Support multiple nvme controllers
In order to support having a "backup" nvme location (for error
handling cases where the device at root port 0 is not found) we
need to be able to support multiple controllers with this driver,
so allocate the cleanup structure instead of using a static variable.

Also clean up the error messages during probing so it is clear what
device is being looked for since they may not be expected to exist.

BUG=b:35581264
TEST=verify logs without any device:
NVMe device not found @ 01:00:00
Updating storage controller failed.
PCIe bridge not found @ 00:1c:04
Updating storage controller failed.

Change-Id: Icd280e1143ac9dec428e890868b83bc11571e4f4
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Reviewed-on: https://chromium-review.googlesource.com/592532
Reviewed-by: Furquan Shaikh <furquan@chromium.org>