TCPC: use EC_CMD_PD_CHIP_INFO for TCPC info
we have a dedicated EC command to get the TCPC version info - no need
to directly scrape the regs.  this can also be used to force the EC to
refresh its view (cache) of the TCPC version after a firmware update,
so this is a good thing all around.

i'm also removing dependencies on PD_SUSPEND and EC I2C tunneling in
the *_check_hash() flow so this can work even when we have
EC_CMD_I2C_PASSTHRU_PROTECT and PD_CONTROL_DISABLE locked down.  this
addresses the case where the AP does a soft reboot without affecting
the EC.

BUG=b:35586896
BRANCH=none
TEST=verified TCPC firmware update works as before.

Change-Id: Iddb60f250bf1ace401114b06733ef02bbe13d9ae
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Commit-Id: e45c8083bd1701cd6f0f979fb49afa64b91bb8d5
Original-Change-Id: Ia64c93bfda50dd5a651836224b9fb9f6ce1db57c
Original-Signed-off-by: Caveh Jalali <caveh@google.com>
Original-Reviewed-on: https://chromium-review.googlesource.com/615089
Original-Reviewed-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/914437