mmc: Remove ERROR string from message
If an eMMC is listed as a possible boot device but is not present
(because another boot device type is) then it will fail to be
detected, which currently prints a string:
ERROR: Card did not respond to voltage select!

This ERROR part of the string makes automated tests that scan for
error messages in the firmware logs fail..

BUG=b:64684441
BRANCH=eve
TEST=boot on eve w/o eMMC and check for this string in the log:
MMC did not respond to voltage select!

Change-Id: I47d4faa008f545d60650be7fde71fcfbf62dea5b
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Commit-Id:
Original-Change-Id: I726fb6d9284f78166fc2027f7e0d2ae6bfcb9c02
Original-Signed-off-by: Duncan Laurie <dlaurie@google.com>
Reviewed-on: https://chromium-review.googlesource.com/658657