crosEC: Do not add padding when writing to flash
When Depthcharge writes to flash through EC, it adds padding to make
the data size equal to write_block_size of the SPI flash. Though this
might be necessary for a certain board (Pit as described in CL:62895),
this fails if we're writing to the end of the SPI flash because EC
checks whether (offset + size) exceeds the flash size or not
(legitimately).

I think we should play this trick (padding to write block size) on the
EC side. That is, if EC receives a short packet, it should add padding
by itself if it's required.

Pit has been removed from the crosEC tree. If this patch breaks another
board, we'll fix it differently (as suggested above).

BUG=chromiumos:756230
BRANCH=none
TEST=Let Depthcharge on Fizz write a RW image to the slot B, which is
located at the end.

Change-Id: I5834836b5de6066a4b8068f0f0fda01f15cf928a
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Commit-Id: 42f2d0c408b2b311886b19d349f8505af81a629d
Original-Change-Id: I2f51917976189ea01de8c5787f6329f0a45c7f74
Original-Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Original-Reviewed-on: https://chromium-review.googlesource.com/618164
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Original-Reviewed-by: Randall Spangler <rspangler@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/914436