anx3429: wait for TCPC restart after FW update
we need to wait for the TCPC to restart and the pd_task() to settle
after a firmware update before proceding.  the AP doesn't really have
a reliable way to to tell if the pd_task() is back in business on the
EC side, so we'll just wait a conservative amount of time before
declaring victory.

BRANCH=none
BUG=b:35586895

TEST=used 2 different firmware blobs and verified it switches between
	them during software sync.

CQ-DEPEND=CL:631976

Change-Id: I983894464cfaa27d5c133c9728c5df11ebb62323
Signed-off-by: Caveh Jalali <caveh@google.com>
Reviewed-on: https://chromium-review.googlesource.com/631996
Reviewed-by: Julius Werner <jwerner@chromium.org>