anx3429: wait for TCPC restart after FW update
we need to wait for the TCPC to restart and the pd_task() to settle
after a firmware update before proceding.  the AP doesn't really have
a reliable way to to tell if the pd_task() is back in business on the
EC side, so we'll just wait a conservative amount of time before
declaring victory.

BRANCH=none
BUG=b:35586895

TEST=used 2 different firmware blobs and verified it switches between
        them during software sync.

CQ-DEPEND=CL:631976

Change-Id: Ief445e439f9afb2d7eec2ee1ddb61c89ca53de47
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Commit-Id: 5b832152e818ca65ebf7732aecb2e9e7031401af
Original-Change-Id: I983894464cfaa27d5c133c9728c5df11ebb62323
Original-Signed-off-by: Caveh Jalali <caveh@google.com>
Original-Reviewed-on: https://chromium-review.googlesource.com/631996
Original-Reviewed-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/914439