screens: Correct text height for WAIT screen
Since bmpblk has been fixed to no longer render an empty line at the end
of the critical update string, this patch corrects the image height
accordingly to keep the text at the same size. (And as a bonus, it's no
longer off-center.)

CQ-DEPEND=CL:639018
BRANCH=none
BUG=none
TEST=Hacked Kevin to display WAIT instead of INSERT, then sanity-checked
the screen in a few languages in recovery mode.

Change-Id: I0cf6143b5a4d746794c611c27d5bbbad9ae23626
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/639238
Reviewed-by: Daisuke Nojiri <dnojiri@chromium.org>