bootdata: Only supply extra header if supported by ramdisk
In order to better support images that do not enable the extra header
in the ramdisk, only add the extra headers if it is present in the
container bootdata for the ramdisk.

BUG=b:38040849
TEST=enable building both kernel and ramdisk with 'extra' header
and ensure it boots properly and can verify CRC.  Also ensure that
images without extra headers still boot properly.

Change-Id: I6a7a439433c9d616886a1857bbc2d0092315c201
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Reviewed-on: https://chromium-review.googlesource.com/689356
Reviewed-by: Furquan Shaikh <furquan@chromium.org>