Add CleanupOnVboot handler
INIT_FUNC handlers (including board_setup) are run before vboot is
initialized, as they may be preparing interfaces for vboot to use.

This means that the board_setup function can't rely on things like
vboot_in_recovery() or access to the GBB flags, as these are not
available until vboot setup has been run.

In order to support this add a cleanup function type that will run
callbacks after vboot is initialized but before control is passed
to VbSelectAndLoadKernel.

This provides a path for board_setup to register a function to be
called before vboot is executed to boot a kernel.

BUG=b:67029862
BRANCH=eve
TEST=manual: boot on Eve with subsequent commit that checks both
GBB flags and vboot_in_recovery().

Change-Id: I75e4d19760bdce936ce49abe88708a6f1e1c8ede
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Reviewed-on: https://chromium-review.googlesource.com/754913
Reviewed-by: Aaron Durbin <adurbin@chromium.org>