eve: Implement workaround for audio
The rt5514 codec is getting into a bad state occasionally, and it
cannot be revived with normal reboot methods because the codec is
on 1.8V DSW power rail which does not drop.

This patch implements a workaround in which we attempt to probe the
codec at boot and issue a reset via cr50 if it is in a bad state.

A byte of unused CMOS is used to ensure the workaround is only
attempted once and the system does not get stuck in a reboot loop.
The workaround is also not attempted in recovery mode and if GBB
flags are set for FAFT testing.

This workaround is done in a "CleanupOnVboot" callback since the
function needs to use both vboot_in_recovery() as well as check GBB
flags, and both of those cannot be done at the INIT_FUNC stage as
vboot has not been initialized yet.

BUG=b:67029862
BRANCH=eve
TEST=manual cold reset testing on 6x Eve units, two of which can
reproduce this failure regularly.  Ensure that the bus is recovered
every time and that audio is always functional in the OS.

Change-Id: I17cf0167610e3f402432be9daac3139df368ff34
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Reviewed-on: https://chromium-review.googlesource.com/754914
Reviewed-by: Aaron Durbin <adurbin@chromium.org>