fit: Redesign compat ranking to work as intended, add SKU ID support
The existing depthcharge code attempts to implement correct FIT
compatible string matching but doesn't really get it right. It assigns
every match a compat_rank based on how many non-matching strings are in
front of it in the FDT currently under consideration. This makes no
sense: it is device tree convention to always put the most specific
compatible strings first and the less specific ones later, so this will
pick less specific matches in an ambiguous situation.

But really, the whole question is sort of pointless anyway: there is no
real ranking between compatible strings in the FDT itself -- if the
string is there, that means this FDT can support it as well as any
other. Where ranking really makes sense is in comparison to the string
we're searching for -- we should first search for more specific strings
of the board we're running on, and if we don't find those we should
search for less specific ones. The current depthcharge code only ever
searched for one string anyway, but looking at the original u-boot
implementation (common/image-fit.c#fit_node_find_compat) we can see
exactly this behavior.

This patch adds support for a SKU ID part in the default compatible
string we're looking for, and uses that opportunity to clean up the
ranking mess as well. Depthcharge can now build a table of increasingly
non-specific compatible strings and search for each of them in a row,
keeping track of the most specific match found for each FDT. In the end,
it selects the FDT that had any match for the most specific string we
searched for. This matches the original u-boot behavior.

BRANCH=none
BUG=b:69379692
TEST=Booted Kevin.

Change-Id: Ia5771f445b0ab10a773a08ab9e899b2e63158c81
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/807627
Reviewed-by: Brian Norris <briannorris@chromium.org>