vboot: Use cleanup for power button inhibiting, restrict to x86+UI
The current code for inhibiting the power button SMI when it is needed
for the detachable UI assumes that VbSelectAndLoadKernel() will always
return. This isn't necessarily the case, for example when using the
legacy boot feature. Therefore, it is better to implement the code
restoring the normal power button behavior in a cleanup handler which
is guaranteed to run on any sort of depthcharge exit (except for an EC
reboot, where it shouldn't matter anyway).

Also restrict the code to x86 boards (which are the only ones that need
this) and to situations that actually have a UI (developer and recovery
mode). This should save a few milliseconds of unnecessary EC
communication in the critical boot path.

BRANCH=None
BUG=None
TEST=None

Change-Id: I87f3c6977a379518c59a81d8acda73a5e4bfa47f
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/812504
Tested-by: Furquan Shaikh <furquan@chromium.org>
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
(cherry picked from commit 0b76a02c5ec9ad9f4a526597114789531bcdb127)
Reviewed-on: https://chromium-review.googlesource.com/817920
Reviewed-by: Shelley Chen <shchen@chromium.org>
Commit-Queue: Shelley Chen <shchen@chromium.org>
Tested-by: Shelley Chen <shchen@chromium.org>