vboot/screens: Reduce divider width to 90%
In order to make them look less cramped on panels with Scarlet-like
aspect ratios, we want to add a bit of padding (5% of screen width on
each side) around the horizontal dividers that frame our firmware
screens. This patch achieves that by lowering the divider width to 90%.

CQ-DEPED=CL:826000
BRANCH=none
BUG=b:70163917
TEST=Boot a Scarlet and admire the perfectly balanced screen while
sipping a grande cappuccino.

Change-Id: Ib7b871c9ea2152497f72426d2db26b35edfeb7b6
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/826682
Reviewed-by: Daisuke Nojiri <dnojiri@chromium.org>