gru: Remove power button and lid switch for Scarlet
Gru-based boards have power button and lid switch routed to the EC. This
means that we need to keep issuing EC_CMD_READ_MEMMAP whenever vboot
polls them. This seems to have been fine for the NPCX, but on Scarlet's
STM32 it seems to be causing starvation issues for the UART that lead to
FAFT failures.

The good news is that Scarlet doesn't really have a lid switch anyway,
and the power button serves a different function for the detachable UI
(where it's handled via MKBP) so vboot doesn't care about it's state for
VbExIsShutdownRequested() anyway. This means we can elegantly swipe the
problem under the rug by just hacking these out.

BRANCH=none
BUG=b:35551132
TEST=Booted Scarlet, observed no more EC command 0x7 spam.

Change-Id: I13eed0eafc7958691c117c59ed1f095e3d06ac8a
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/826469
Tested-by: Philip Chen <philipchen@chromium.org>
Reviewed-by: Philip Chen <philipchen@chromium.org>
Reviewed-by: Shawn N <shawnn@chromium.org>