vboot/screens: Fix Noto Sans arrows and reduce text space padding
The recent switch from Noto Sans UI to Noto Sans had a weird effect on
the arrow characters that we display next to the language -- they are no
longer vertically centered on the text line. This patch applies a little
offset to them to fix the problem.

Also reduce the "padding" that we insert to represent a normal text
space between two individually rendered pieces of text that should form
a continuous sentence, because I really feel like the current amount
looks far too wide. (This may have also been affected by the font, since
it seems to already render some amount of white space on the sides of an
image representing text.)

BRANCH=none
BUG=none
TEST=Looked through the screens for all languages on Kevin.

Change-Id: I7168705406472f986a4b81b45fd1316868fa78c7
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/826683
Reviewed-by: Daisuke Nojiri <dnojiri@chromium.org>